Update README.md
Conduct an analysis of your choice either on your mobile phone, mobile device, laptop, or desktop. Some examples of analysis you may want to conduct include the iphone analyzer, digital forensics framework, and or utilize The Sleuth kit if using a Unix system. Submit screenshots of your analysis along with a one page typed report in APA format that includes a description of the tool used and a discussion of your overall analysis.